+++ tests/test_streaming.py
@@ -2,8 +2,8 @@ import tda
 import urllib.parse
 import json
 import copy
-from ._utils import account_principals, has_diff, MockResponse
-from ._utils import no_duplicates
+from .utils import account_principals, has_diff, MockResponse
+from .utils import no_duplicates
 from unittest import IsolatedAsyncioTestCase
 from unittest.mock import ANY, AsyncMock, call, MagicMock, Mock, patch
 from tda import streaming
