+++ tests/test_utils.py
@@ -8,7 +8,7 @@ from tda.orders import EquityOrderBuilder
 from tda.utils import AccountIdMismatchException, Utils
 from tda.utils import UnsuccessfulOrderException
 from . import test_utils
-from ._utils import no_duplicates, MockResponse
+from .utils import no_duplicates, MockResponse
 
 
 class UtilsTest(unittest.TestCase):
